requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
bcrypt
python-dotenv

#Για να τα τρεξεις κανεις στο τερμιναλ pip install -r requirements.txt